test: move tests into same package as production code
Convert package X_test → package X for snapshots_test.go, node_test.go,
joined_test.go, and pprof_test.go. Drops the self-imports and strips
package qualifiers from referenced symbols. Aligns these files with the
repo convention of co-locating tests with the code they exercise.
